cleverman/requirements.txt
astroid==2.3.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
connexion==2.7.0
daiquiri==2.1.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-spec-validator==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyrsistent==0.16.0
python-json-logger==0.1.11
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
swagger-ui-bundle==0.0.6
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
